+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Fix error handling in MastoApi.getServerSideFilters()
 
 ### v0.69.21
 * Fix filtering of user's own toots out of their timeline